[Hyde]: krol430 updated backups/creating-backups.md
In a perfect world, this would mean that your `Thesis.docx` exists on your computer, on another computer or external backup drive,  and on a cloud drive (Google Drive, OneDrive etc). How far you want to go with this rule is part of your planning.In a perfect world, this would mean that your `Thesis.docx` exists on your computer, on another computer or external backup drive,  and on a cloud drive (Google Drive, OneDrive etc). How far you want to go with this rule is part of your planning.asdfasdf